+++ database/migrations/2024_05_12_162919_create_line_items_table.php
@@ -33,7 +33,7 @@ return new class extends Migration
             $table->softDeletes();
             $table->timestamps();
 
-            $table->foreignId('order_id')->constrained();
+            $table->foreignId('order_id')->constrained()->onDelete('cascade');
         });
     }