+++ routes/v1/api/order.php
@@ -11,4 +11,6 @@ use Illuminate\Support\Facades\Route;
 
 Route::prefix('orders')->group(function () {
   Route::get('', [OrderController::class, 'index'])->name('orders.index');
+  Route::get('woo',[OrderController::class,'getOrderWoo'])->name('orders.woo-orders');
+  Route::get('syncOrder',[OrderController::class,'syncNewOrders'])->name('orders.syncOrders');
 });